modify to use float

+++ modules/quadramp/quadramp.h
@@ -39,14 +39,14 @@
  */
 struct quadramp_filter
 {
-	double var_2nd_ord_pos; /**< Acceleration */
-	double var_2nd_ord_neg; /**< Deceleration */
-	double var_1st_ord_pos; /**< Speed (> 0) */
-	double var_1st_ord_neg; /**< Speed (< 0) */
+	float var_2nd_ord_pos; /**< Acceleration */
+	float var_2nd_ord_neg; /**< Deceleration */
+	float var_1st_ord_pos; /**< Speed (> 0) */
+	float var_1st_ord_neg; /**< Speed (< 0) */
 
-	double previous_var; /**< Speed at the previous filter iteration. */
-	double previous_out; /**< Position at the previous filter iteration. */
-	int32_t previous_in; /**< Input at the previous filter iteration. */
+	float previous_var; /**< Speed at the previous filter iteration. */
+	float previous_out; /**< Position at the previous filter iteration. */
+	float previous_in; /**< Input at the previous filter iteration. */
 };
 
 /** Initialization of the filter.
@@ -65,8 +65,8 @@ void quadramp_reset(struct quadramp_filter *q);
  * @param [in] var_2nd_ord_neg The negative acceleration.
  */
 void quadramp_set_2nd_order_vars(struct quadramp_filter *q,
-				 double var_2nd_ord_pos,
-				 double var_2nd_ord_neg);
+				 float var_2nd_ord_pos,
+				 float var_2nd_ord_neg);
 
 /**@brief Set speed. 
  * @param [in] q The quadramp instance.
@@ -74,8 +74,8 @@ void quadramp_set_2nd_order_vars(struct quadramp_filter *q,
  * @param [in] var_1st_ord_neg The negative speed.
  */
 void quadramp_set_1st_order_vars(struct quadramp_filter *q,
-				 double var_1st_ord_pos,
-				 double var_1st_ord_neg);
+				 float var_1st_ord_pos,
+				 float var_1st_ord_neg);
 
 /** @brief Set position.
  *
@@ -83,7 +83,7 @@ void quadramp_set_1st_order_vars(struct quadramp_filter *q,
  * @param [in] q The quadramp instance.
  * @param [in] pos The new position.
  */
-void quadramp_set_position(struct quadramp_filter *q, int32_t pos);
+void quadramp_set_position(struct quadramp_filter *q, float pos);
 
 /** @brief Is the ramp finished.
  *
@@ -99,6 +99,6 @@ uint8_t quadramp_is_finished(struct quadramp_filter *q);
  *
  * @returns The output of the filter.
  */
-int32_t quadramp_do_filter(void *data, int32_t in);
+float quadramp_do_filter(void *data, float in);
 
 #endif
